print list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,10 +8,10 @@
 
 size_t print_list(const list_t *h)
 {
-	unsigned int n = 1;
+	unsigned long  int n = 1;
 
 	if (h->str)
-		printf("[%d] %s\n", h->len, h->str);
+		printf("[%u] %s\n", h->len, h->str);
 	else
 		printf("[0] (nil)\n");
